Image-Processing-&-Object-Recognition-Pipeline/requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# ML / Computer Vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
sahi>=0.11.14
transformers>=4.30.0
timm>=0.9.0
einops>=0.6.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Utilities
scikit-learn>=1.3.0
huggingface_hub>=0.16.0

# Google Gemini
google-genai>=0.3.0

# Vector Search (FAISS)
faiss-cpu>=1.7.4

# Image Enhancement (SwinIR)
basicsr>=1.4.2

# Database & Cache
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0

# Configuration
pydantic-settings
